+++ connector_fennoa/components/__init__.py
@@ -1,0 +1,1 @@
+from . import fennoa_export_mapper