+++ connector_fennoa/readme/HISTORY.md
@@ -1,3 +1,23 @@
+17.0.1.2.0 (2026-03-21)
+
+Add import/export mappers
+
+17.0.1.1.1 (2026-02-19)
+
+Add shipping address information
+
+17.0.1.1.0 (2026-02-19)
+
+Add sale invoice approve/sending configuration.
+
+17.0.1.0.2 (2026-02-13)
+
+Improve field mapping
+
+17.0.1.0.1 (2026-01-07)
+
+Production version
+
 17.0.0.1.0 (2025-12-08)
 
 Initial version